cloud/docker_tutorials/micro_services/spell_checker/requirements.txt
pyenchant==3.1.1
Flask~=1.1.2